Split pgtoolkit.log module
