+++ pgtoolkit/log/parser.py
@@ -1,117 +1,6 @@
 # coding: utf-8
-"""\
-.. currentmodule:: pgtoolkit.log
-
-Postgres logs are still the most comprehensive source of information on what's
-going on in a cluster. :mod:`pgtoolkit.log` provides a parser to exploit
-efficiently Postgres log records from Python.
-
-Parsing logs is tricky because format vary accros configurations. Also
-performance is important while logs can contain thousands of records.
-
-
-Configuration
--------------
-
-Postgres log records have a prefix, configured with ``log_line_prefix`` cluster
-setting. When analyzing a log file, you must known the ``log_line_prefix``
-value used to generate the records.
-
-Postgres can emit more message for your needs. See `Error Reporting and Logging
-section
-<https://www.postgresql.org/docs/current/static/runtime-config-logging.html>`_
-if PostgreSQL documentation for details on logging fields and message type.
-
-
-Performance
------------
-
-The fastest code is NOOP. Thus, the parser allows you to filter records as soon
-as possible. The parser has several distinct stages. After each stage, the
-parser calls a filter to determine whether to stop record processing. Here are
-the stages in processing order :
-
-1. Split prefix, severity and message, determine message type.
-2. Extract and decode prefix data
-3. Extract and decode message data.
-
-
-Limitations
------------
-
-:mod:`pgtoolkit.log` does not manage opening and uncompressing logs. It only
-accepts a line reader iterator that loops log lines. The same way,
-:mod:`pgtoolkit.log` does not manage to start analyze at a specific position in
-a file.
-
-:mod:`pgtoolkit.log` does not gather record set such as ``ERROR`` and
-following ``HINT`` record. It's up to the application to make sense of record
-sequences.
-
-:mod:`pgtoolkit.log` does not analyze log records. It's just a parser, a
-building block to write a log analyzer in your app.
-
-
-API Reference
--------------
-
-Here are the few functions and classes used to parse and access log records.
-
-.. autofunction:: parse
-.. autoclass:: Record
-.. autoclass:: UnknownData
-.. autoclass:: NoopFilters
-
-
-Example
--------
-
-Here is a sample structure of code parsing a plain log file.
-
-.. code-block:: python
-
-    with open('postgresql.log') as fo:
-        for r in parse(fo, prefix_fmt='%m [%p]'):
-        if isinstance(r, UnknownData):
-            "Process unknown data"
-        else:
-            "Process record"
-
-
-
-Using :mod:`pgtoolkit.log` as a script
---------------------------------------
-
-You can use this module to dump logs as JSON using the following usage::
-
-    python -m pgtoolkit.log <log_line_prefix> [<filename>]
-
-:mod:`pgtoolkit.log` serializes each record as a JSON object on a single line.
-
-.. code:: console
-
-    $ python -m pgtoolkit.log '%m [%p]: [%l-1] app=%a,db=%d%q,client=%h,user=%u ' data/postgresql.log
-    {"severity": "LOG", "timestamp": "2018-06-15T10:49:31.000144", "message_type": "connection", "line_num": 2, "remote_host": "[local]", "application": "[unknown]", "user": "postgres", "message": "connection authorized: user=postgres database=postgres", "database": "postgres", "pid": 8423}
-    {"severity": "LOG", "timestamp": "2018-06-15T10:49:34.000172", "message_type": "connection", "line_num": 1, "remote_host": "[local]", "application": "[unknown]", "user": "[unknown]", "message": "connection received: host=[local]", "database": "[unknown]", "pid": 8424}
-
-"""  # noqa
-
-from __future__ import print_function
-
-from datetime import datetime
-import json
-import logging
-import os
 import re
-import sys
-from datetime import timedelta
-
-from ._helpers import JSONDateEncoder
-from ._helpers import open_or_stdin
-from ._helpers import Timer
-
-
-logger = logging.getLogger(__name__)
+from datetime import datetime, timedelta
 
 
 def parse(fo, prefix_fmt, filters=None):
@@ -572,33 +461,3 @@ class Record(object):
             (k, v)
             for k, v in self.__dict__.items()
         ])
-
-
-def main(argv=sys.argv[1:], environ=os.environ):
-    logging.basicConfig(
-        level=logging.INFO,
-        format='%(asctime)s %(levelname)5.5s %(message)s',
-    )
-
-    argv = argv + ['-']
-    log_line_prefix = argv[0]
-    counter = 0
-    try:
-        with open_or_stdin(argv[1]) as fo:
-            with Timer() as timer:
-                for record in parse(fo, prefix_fmt=log_line_prefix):
-                    if isinstance(record, UnknownData):
-                        logger.warning("%s", record)
-                    else:
-                        counter += 1
-                        print(
-                            json.dumps(record.as_dict(), cls=JSONDateEncoder))
-        logger.info("Parsed %d records in %s.", counter, timer.delta)
-    except Exception as e:
-        print(str(e), file=sys.stderr)
-        return 1
-    return 0
-
-
-if '__main__' == __name__:  # pragma: nocover
-    sys.exit(main(argv=sys.argv[1:], environ=os.environ))
